engine: extract esm/folder.js
